"fix(validation): add conditional validation for user type"

+++ src/main/java/proyecto/spring/asugestionsocios/model/dto/PaymentDTO/MonthlyFeeDTO.java
@@ -5,6 +5,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -14,10 +15,9 @@ public class MonthlyFeeDTO {
     @Positive(message = "Monthly fee number must be a positive value")
     private Integer monthlyFeeNumber;
 
-    //TODO: Changes data type in database to BigDecimal
     @NotNull(message = "Total annual monthly fee is mandatory")
     @Positive(message = "Total annual monthly fee must be a positive value")
-    private Integer totalAnnualMonthlyFee;
+    private BigDecimal totalAnnualMonthlyFee;
 
     @NotNull(message = "Payment due date is mandatory")
     @Future(message = "Payment due date must be in the future")
